Test io-file on TravisCI

+++ tests/io-file/test.cpp
@@ -237,7 +237,7 @@ void test_write_read ()
     TEST_FAIL((d = open_device(open_params<file>(file_path, pfs::io::read_only))));
     pfs::byte_string bs;
     std::cout << "d.available()=" << d.available() << std::endl;
-    d.read(bs, d.available());
+    d.read(bs, static_cast<size_t>(d.available()));
 
     std::cout << "bs.size()=" << bs.size() << std::endl;
     TEST_OK(d.close());
